Close websocket connection asynchronously
This fixes a hang that would happen on shutdown which was being caused by the websocket taking a short amount of time to fully disconnect. This may not be 100% correct, but the server should also be able to detect a disconnected client via a timeout, so in practice it shouldn't be an issue.